ChallengeGoldLevel/requirements.txt
Flask==2.1.3
Flasgger==0.9.5
pandas==1.5.3
matplotlib==3.6.2
seaborn==0.12.1
pyyaml==6.0
sqlite3==3.36.0
